requirements.txt
audioread==2.1.8
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.18.2
idna==2.8
joblib==0.14.1
librosa==0.7.2
llvmlite==0.31.0
numba==0.48.0
numpy==1.18.1
Pillow==7.0.0
pycparser==2.19
python-telegram-bot==12.3.0
requests==2.22.0
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
tornado==6.0.3
urllib3==1.25.8